+++ src/main/kotlin/team/mozu/dsm/domain/lesson/model/id/LessonItemId.kt
@@ -1,0 +1,8 @@
+package team.mozu.dsm.domain.lesson.model.id
+
+import java.util.UUID
+
+data class LessonItemId(
+    val lessonId: UUID,
+    val itemId: UUID
+)
